moving serve to rfosc_server
